+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Suivi des changements de version par classement antéchronologique.
 
-## ST-2.0.0
+## ST-2.0.0 - 2026-09-26
 
 ### Ajouts
 
@@ -10,6 +10,7 @@ Suivi des changements de version par classement antéchronologique.
 1. contenu: gestion des images locales incluses dans les markdown
 1. **navigation** : nombre de mots du post et temps de lecture
 1. navigation : date de mise à jour du post
+1. **site** : pages de contenu markdown et d'images fournies en exemple
 
 ## ST-1.1.0 - 2026-05-08